requirements.txt
flask>=2.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
